finish background music and black
